Add ignore transaction utility
Move transaction utilities to their own file.
Cleanup PR template.
