+++ edx_django_utils/__init__.py
@@ -2,7 +2,7 @@
 EdX utilities for Django Application development..
 """
 
-__version__ = "3.5.0"
+__version__ = "3.6.0"
 
 default_app_config = (
     "edx_django_utils.apps.EdxDjangoUtilsConfig"
