+++ edx_django_utils/monitoring/__init__.py
@@ -3,13 +3,5 @@ Metrics utilities public api
 
 See README.rst for details.
 """
-
-from .utils import (
-    accumulate,
-    function_trace,
-    get_current_transaction,
-    increment,
-    set_custom_metric,
-    set_custom_metrics_for_course_key,
-    set_monitoring_transaction_name
-)
+from .transactions import function_trace, get_current_transaction, ignore_transaction, set_monitoring_transaction_name
+from .utils import accumulate, increment, set_custom_metric, set_custom_metrics_for_course_key
